Create file for initial state
